cleanup: drop duplicate rag-wrapper libs rag-wrapper (an internal library) was defined in 3 places - in components/, in shiksha-api/app-service, and in shiksha-api/durable-functions. From what I gather, components/ was meant to address the boilerplate, but whoever worked on that forgot to complete this work.
