--- shiksha-api/app-service/app/rag-wrapper/.env.example
@@ -1,23 +1,0 @@
-# Azure OpenAI Configuration (Required)
-AZURE_OPENAI_API_KEY=your_azure_openai_api_key_here
-AZURE_OPENAI_ENDPOINT=https://your-openai-resource.openai.azure.com/
-AZURE_OPENAI_EMBEDDING_MODEL=text-embedding-ada-002
-AZURE_OPENAI_EMBEDDING_DEPLOYMENT=text-embedding-ada-002
-AZURE_OPENAI_COMPLETION_MODEL=gpt-35-turbo
-AZURE_OPENAI_COMPLETION_DEPLOYMENT=gpt-35-turbo
-AZURE_OPENAI_API_VERSION=2024-02-15-preview
-
-# Neo4j Configuration (Required for Neo4j tests)
-NEO4J_URI=bolt://localhost:7687
-NEO4J_USER=neo4j
-NEO4J_PASSWORD=password
-
-# Qdrant Configuration (Required for Qdrant tests)
-QDRANT_URL=http://localhost:6333
-QDRANT_API_KEY=your_qdrant_api_key_here_if_needed
-QDRANT_COLLECTION_NAME=test_neo4j_qdrant
-
-# Azure AI Search Configuration (Required for Azure AI Search tests)
-AZURE_AI_SEARCH_ENDPOINT=https://your-search-service.search.windows.net
-AZURE_AI_SEARCH_API_KEY=your_azure_ai_search_api_key_here
-AZURE_AI_SEARCH_INDEX_NAME=test-index
