--- shiksha-api/app-service/app/rag-wrapper/.python-version
@@ -1,1 +1,0 @@
-3.11.13
