+++ shiksha-api/app-service/docker_commands.txt
@@ -1,4 +1,4 @@
 az acr login --name srivarsiregistry
-docker build -t srivarsiregistry.azurecr.io/shikshafastapi:latest .
+docker build -f shiksha-api/app-service/Dockerfile -t srivarsiregistry.azurecr.io/shikshafastapi:latest .
 docker push srivarsiregistry.azurecr.io/shikshafastapi:latest
 docker system prune -a --volumes
