--- shiksha-api/durable-functions/core/rag-wrapper/.gitignore
@@ -1,3 +1,0 @@
-**/__pycache__/*
-**/.env
-tests/indexes/
