--- shiksha-api/durable-functions/core/rag-wrapper/.python-version
@@ -1,1 +1,0 @@
-3.11.13
