+++ shiksha-api/durable-functions/pyproject.toml
@@ -17,7 +17,7 @@ openai = "1.81.0"
 azure-storage-blob = "^12.25.1"
 aiofiles = "^24.1.0"
 azure-identity = "^1.23.0"
-vellm-rag-wrapper = { path = "core/rag-wrapper", develop = true }
+vellm-rag-wrapper = { path = "../../components/rag-wrapper", develop = true }
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.0.0"
